Move git package to libs/git
