+++ libs/git/ignore.go
