+++ libs/git/testdata/.gitignore
