+++ libs/git/testdata/a/.gitignore
