+++ libs/git/testdata/a/b/.gitignore
